Fastforward Public Develop (#45)
* 3D group forward convolution solver (#2286)

* [HotFix] Fix build issue after #2286 (#2328)

* Fix ConvCkIgemmFwdV6r1DlopsNchw solver to reflect that it's not dynamic (#2325)

* Remove target IDs from kdb entries (#2309)

* Dropout: make seed and states_num kernel arguments (#2277)

* [MI250] Adding missing kernel objects (#2329)

* Post-merge fixups: Replace environment variable check with problem config check and reduce lambda capture for Invoker obj (#2305)

* [HotFix][CI] fix HIP tidy issue from #2277 (#2335)

* [HotFix] Update requirements.txt MLIR ignore PATH for Python conda LLD (#2324)

* [NFC] Replace long integers with int64_t and size_t for better compatibility with Windows (#2323)

* Windows compatibility: replace long integers with int64_t and size_t, replace uint with unsigned int, replace long with long long for numbers, proper casting

* Fix formatting

* Fix 3d group forward convolution

* Resolve review comments

* Fix formatting

---------

+++ test/tensor_holder.hpp
@@ -520,7 +520,7 @@ void generate_unary_one(F f, std::vector<int> input, G g)
 
 struct tensor_elem_gen_integer
 {
-    unsigned long max_value = 17;
+    uint64_t max_value = 17;
 
     template <class... Ts>
     double operator()(Ts... Xs) const
@@ -528,9 +528,10 @@ struct tensor_elem_gen_integer
         static_assert(sizeof...(Ts) < 6,
                       "Dimensions in tensor_elem_gen_integer must be less than 6.");
         assert(max_value > 0);
-        std::array<unsigned long, sizeof...(Ts)> left = {{Xs...}};
-        std::array<unsigned long, 5> right            = {{613, 547, 701, 877, 1049}};
-        unsigned long dot = std::inner_product(left.begin(), left.end(), right.begin(), 173ul);
+        std::array<uint64_t, sizeof...(Ts)> left = {{Xs...}};
+        std::array<uint64_t, 5> right            = {{613, 547, 701, 877, 1049}};
+        uint64_t dot =
+            std::inner_product(left.begin(), left.end(), right.begin(), static_cast<uint64_t>(173));
         return static_cast<double>(dot % max_value);
     }
 };
